- Split off registration-centric Jira services into `Otis::JiraClient::Registration` subclass. - Inform user whether a Jira ticket was created or updated.
